typo (COLOR) fix some comments and typos

+++ viz/d2v-viz-jfx/src/main/kotlin/io/data2viz/viz/JFxVizRenderer.kt
@@ -183,6 +183,6 @@ fun LineNode.render(renderer: JFxVizRenderer){
 }
 
 val Color.jfxColor: javafx.scene.paint.Color
-    get() = javafx.scene.paint.Color.rgb(r, g, b, alpha.toDouble())
+    get() = javafx.scene.paint.Color.rgb(r, g, b, alpha)
